feat: watchOS와 iOS 분리
