+++ Projects/App/WatchExtension/Sources/WatchApp.swift
@@ -11,15 +11,9 @@ import Presentation
 
 @main
 struct WatchApp: App {
-    let compositionRoot: CompositionRootInterface
-    
-    public init() {
-        self.compositionRoot = CompositionRoot()
-    }
     @SceneBuilder var body: some Scene {
         WindowGroup {
             IntervalScreen(intervalDIContainer: IntervalDIContainer())
-//            WatchIntervalScreen(watchIntervalDIContainer: compositionRoot.intervalDIContainer)
         }
     }
 }
