+++ Projects/Presentation/Sources/Interval/Active/Screen/WatchOS/IntervalActiveScreen+WatchOS.swift
@@ -1,0 +1,15 @@
+//
+//  IntervalActiveScreen+WatchOS.swift
+//  Presentation
+//
+//  Created by 송영모 on 12/3/23.
+//
+
+import Foundation
+import SwiftUI
+
+public extension IntervalActiveScreen {
+    var watchOS: some View {
+        IntervalActiveWatchScreen(viewModel: viewModel)
+    }
+}
