+++ Projects/Presentation/Sources/Interval/Active/Screen/iOS/IntervalActiveIPhoneScreen.swift
@@ -1,0 +1,17 @@
+//
+//  IntervalActiveScreenWithiPhone.swift
+//  Presentation
+//
+//  Created by 송영모 on 12/3/23.
+//
+
+import Foundation
+import SwiftUI
+
+public struct IntervalActiveIPhoneScreen: View {
+    @ObservedObject var viewModel: IntervalActiveViewModel
+    
+    public var body: some View {
+        Text("여기는 아이폰")
+    }
+}
