+++ Projects/Presentation/Sources/Interval/Active/ViewModel/IntervalActiveViewModel+IntervalRouter.swift
@@ -25,4 +25,3 @@ public class IntervalActiveViewModelWithRouter: IntervalActiveViewModel {
         router.triggerScreenTransition(route: .intervalActive)
     }
 }
-
