+++ Projects/Presentation/Sources/Interval/List/Screen/Cell/IntervalInfoCellView.swift
